Fix a reserved identifier that my scripts missed
My scripts missed it because it was in eventdns.c, which was in ext,
but it _was_ using one of our identifiers.  That's probably because
eventdns.c has drifted a bit since we forked it.

I'm not going to fix the other reserved identifiers in eventdns.c,
since that would make it drift even more.