Make a distinction between asking for loopback interface and otherwise
We want to have an API that specifically allows
retrieving address of a local loopback interface. On
the other hand, we also want to have a way to ask for
interface that is guarranteed to be suitable for public
communications. Platform-specific get_interface_address*
functions are augmented with loopback argument to specify
which kind of interface we want. Furthermore, we introduce
get_loopback_address6{_list} to let the rest of Tor codebase
ask for a (list of) loopback interface addresses. Lastly,
we make tweaks to network interface lookup code to filter
out non-suitable interfaces more efficiently.