rust: Fix a difference in compute_vote() behaviour to C version.
The C version of compute_vote() calls expand_protocol_list() which checks if
there would be too many subprotocols *or* expanded individual version numbers,
i.e. more than MAX_PROTOCOLS_TO_EXPAND, and does this *per vote* (but only in
compute_vote(), everywhere else in the C seems to only care about the number of
subprotocols, not the number of individual versions).  We need to match its
behaviour in Rust and ensure we're not allowing more than it would to get the
votes to match.