main: Set up periodic events by roles only
This is the next step where we only setup the events for the roles we are
configured for.

This commit builds but won't work because we can setup twice the same events
because the roles can be combined. Next commit will add a fix to guard against
that since now it is a reality.

Signed-off-by: David Goulet <dgoulet@torproject.org>