Allow FQDN in config file DA "address" field
Currently, src/or/config.c:parse_dir_authority_line() hard-codes the
weak assumption that the address component of a DirAuthority entry
is a numeric dotted-quad IP address, and generates an error on most
FQDN hostnames. The assumption is weak because if the FQDN begins
with a digit (e.g., "2byfour.example.net"), parsing succeeds, and
the FQDN is successfully resolved later on by addr_port_lookup().

This patch explicitly permits any valid FQDN to be used for a DA's
address by reserving the last two items in the smartlist for the
"address:port" and "fingerprint" fields, respectively.