Allow FQDN in "dir-address" field of DA certificate
Instead of immediately resolving "address" to an IP before
publishing it as part of the "dir-address" field of the new
certificate, make tor-gencert publish it as-is and leave
FQDN-to-IP resolution up to the recipient(s).

This patch also modifies the corresponding parsing routine in
src/or/routerparse.c:authority_cert_parse_from_string() to
resolve the address field at run time from either dotted-quad
or FQDN format.