hs-v3: Implement descriptor replaycache for HSDir
In an effort to remove the descriptor revision counter, we now use a
replaycache at the directory which keeps track of descriptors it sees for a
period of 36h.

A hash of the descriptor _without_ the signature is used in order to avoid a
possible successful replay attack by exploiting the possible signature
malleability of the descriptor.

Now, the reason why we do want to remove the revision counter is because we
want to support the scalaibility use case where multiple services with the
same onion address compete to upload their descriptors. With the revision
counter, if one service would restart more frequently for whatever reasons,
let say maintanance, its revision counter would grow faster and thus it would
become the only service able to upload a new descriptor. This is not good.

Fixes #25552

Signed-off-by: David Goulet <dgoulet@torproject.org>