fixup! Ticket #25573: Check half-opened stream ids when choosing a new one
Unlike the rest of this branch, this commit does change client behavior and
also avoids data corrupton by avoiding mixing up old stream ids with new ones.
Does that make it a feature or a bug? Who can say.

This is technically a feature change from like the first version of Tor also.