fix rust dependency updating instructions
None of the instructions mentioned the key step of updating
Cargo.lock, which the script updateRustDependencies.sh didn't do,
because it used `--locked`. Unless Cargo.lock was already updated
to newer dependencies, the script refuses to do anything at all.

Make sure the commit that updates Cargo.lock is the same commit
that updates the hash for the git submodule.