Ticket #25573: Sort+bsearch the half-closed connection list
It should not be the case that this list gets large enough to matter, but
sorting it also makes it easier to ensure against duplicate elements.

A sorted list is preferable to a hashtable because this set should be very
small. There should only be a handful of half-closed connections on a circuit,
if that. This makes the memory allocation, memory fragmentation, and hashing
overhead of the hash table not worth it.