Try to improve clarity of msg.h
Add a long comment about how to use it, document its internals, and
expand on the existing documentation.

Also, try to disambiguate "ctype" -- use "consttype" when we mean
the constant version of a type, and "c_type" when we mean a type in
C, as opposed to the unique identifier ("typename") for a type.